feat: bundle perf (wip)
