+++ apps/codeimage/src/App.tsx
@@ -1,37 +1,14 @@
-import {Canvas} from './components/Scaffold/Canvas/Canvas';
-import {Toolbar} from './components/Toolbar/Toolbar';
-import {Sidebar} from './components/Scaffold/Sidebar/Sidebar';
-import {createEffect, createSignal, lazy, on, Show, Suspense} from 'solid-js';
-import {Footer} from './components/Footer/Footer';
-import {createI18nContext, useI18n} from '@codeimage/locale';
+import {createEffect, createSignal, lazy, on} from 'solid-js';
+import {useI18n} from '@codeimage/locale';
 import {useModality} from './core/hooks/isMobile';
 import {NotificationHandler} from './components/ui/Toast/SnackbarHost';
 import {PortalHost} from './components/ui/PortalHost/PortalHost';
-import {KeyboardShortcuts} from './components/KeyboardShortcuts/KeyboardShortcuts';
-import {Box} from './components/ui/Box/Box';
 import {initEffects, registerEffects} from '@ngneat/effects';
 import {onTabNameChange$, onThemeChange$} from '@codeimage/store/effect';
 import {uiStore} from './state/ui';
-import {locale} from './i18n';
 
 initEffects();
 registerEffects([onTabNameChange$, onThemeChange$]);
-
-const EditorHandler = lazy(
-  () => import('./components/CustomEditor/EditorHandler'),
-);
-
-const BottomBar = lazy(() => {
-  return import('./components/BottomBar/BottomBar').then(e => {
-    return {default: e.BottomBar};
-  });
-});
-
-const EditorSidebar = lazy(() => {
-  return import('./components/LeftSidebar/EditorSidebar').then(e => ({
-    default: e.EditorSidebar,
-  }));
-});
 
 const ThemeSwitcher = lazy(() => {
   return import('./components/ThemeSwitcher/ThemeSwitcher').then(e => ({
@@ -43,11 +20,14 @@ const ReloadPrompt = lazy(() => {
   return import('./components/PromptUpdate/PromptUpdate');
 });
 
+const Content = lazy(() => {
+  const modality = useModality();
+  return modality === 'full' ? import('./Desktop') : import('./Mobile');
+});
+
 const App = () => {
   document.querySelector('#launcher')?.remove();
-  const [frameRef, setFrameRef] = createSignal<HTMLElement>();
-  const [portalHostRef, setPortalHostRef] = createSignal<HTMLElement>();
-  const modality = useModality();
+  const [, setPortalHostRef] = createSignal<HTMLElement>();
   const [, {locale}] = useI18n();
 
   createEffect(on(() => uiStore.locale, locale));
@@ -55,48 +35,9 @@ const App = () => {
   return (
     <>
       <NotificationHandler />
-
-      <Suspense>
-        <ReloadPrompt />
-      </Suspense>
-
-      <Show when={modality === 'full'}>
-        <Sidebar>
-          <Suspense>
-            <EditorSidebar />
-          </Suspense>
-        </Sidebar>
-      </Show>
-
+      <ReloadPrompt />
       <PortalHost ref={setPortalHostRef} />
-
-      <Canvas>
-        <Toolbar canvasRef={frameRef()} />
-
-        <Show when={modality === 'full'}>
-          <Box paddingLeft={'4'} paddingTop={'3'}>
-            <KeyboardShortcuts />
-          </Box>
-        </Show>
-
-        <Suspense>
-          <EditorHandler frameRef={setFrameRef} />
-        </Suspense>
-
-        <Footer />
-      </Canvas>
-
-      {modality === 'mobile' ? (
-        <Suspense>
-          <BottomBar portalHostRef={portalHostRef()} />
-        </Suspense>
-      ) : (
-        <Sidebar>
-          <Suspense>
-            <ThemeSwitcher orientation={'vertical'} />
-          </Suspense>
-        </Sidebar>
-      )}
+      <Content />
     </>
   );
 };
